Switch to Eldev
...which (a) can install updated versions of built-in packages (seq in
this case) and (b) allows me to at least write some hacky
configuration to get it to load the new version of seq.

It also has a quick way to test on other Emacs versions utilizing
Docker, which is much nicer than the "push to GitHub, wait for CI,
look at the message" iteration cycle.